requirements.txt
langchain
langchain-core
langchain-ollama
langchain-google-genai
google-generativeai
python-dotenv
streamlit
pdfplumber
PyPDF2
pytesseract
Pillow